이력 조회 + 페이징

+++ src/main/java/com/codehows/ksisbe/history/dto/CrawlWorkHistory.java
@@ -17,7 +17,7 @@ public class CrawlWorkHistory {
     private String settingName;
     private Long scheduleId;
     private Long startedBy;
-    private Long failCount;
+    private Integer failCount;
     private String state;
     private String type;
     private LocalDateTime startAt;
